modules added  and linked to index.js
