+++ index.js
@@ -1,25 +1,19 @@
+import { startTime, checkTime } from './modules/Time';
+import Libro from './modules/books.js';
+import printList from './modules/cardhtml.js';
+
 const booksBtn = document.getElementById('booksBtn');
 const addBookbtn = document.getElementById('addbookBtn');
 const contactBtn = document.getElementById('contactBtn');
 const containerBooks = document.getElementById('container');
-const books = document.getElementById('books');
+// const books = document.getElementById('books');
 const addNew = document.getElementById('addbook');
 const contact = document.getElementById('contact');
 const titulo = document.getElementById('titulo');
 const autor = document.getElementById('autor');
 const addBtn = document.getElementById('btn');
-let dataBooks = JSON.parse(localStorage.getItem('book')) || [];
-
-// This code below was recovered from https://programacion.net/
-
-// This code above was recovered from https://programacion.net/
 
 window.onload = startTime();
-
-
-
-
-
 printList();
 
 addBtn.addEventListener('click', () => {
