+++ modules/cardhtml.js
@@ -1,0 +1,27 @@
+import Libro from './books.js';
+
+const books = document.getElementById('books');
+let dataBooks = JSON.parse(localStorage.getItem('book')) || [];
+
+const printList = () => {
+  const libro = new Libro();
+  books.innerHTML = null;
+  if (localStorage.getItem('book')) {
+    dataBooks = JSON.parse(localStorage.getItem('book')) || [];
+  }
+  for (let i = 0; i < dataBooks.length; i += 1) {
+    const bookStorage = document.createElement('div');
+    const bookInfo = document.createElement('p');
+    const deleteBtn = document.createElement('button');
+    bookInfo.textContent = `"${dataBooks[i].title}" by ${dataBooks[i].author}`;
+    deleteBtn.textContent = 'Remove';
+    books.appendChild(bookStorage);
+    bookStorage.append(bookInfo, deleteBtn);
+    bookStorage.className = 'bookStorage';
+    deleteBtn.addEventListener('click', () => {
+      bookStorage.remove();
+      libro.deleteBook(i);
+    });
+  }
+};
+export default { printList, books };
